Refactor match logic in to util
